Fix build error
:Release Notes:
Fix build error

:Detailed Notes:
Build Failed
http://gecko.lge.com/jenkins/job/webos-master-verify-raspberrypi4/12201/ : FAILURE
http://gecko.lge.com/jenkins/job/webos-master-verify-raspberrypi4-64/12149/ : SUCCESS

:Testing Performed:
Local test : OK

:QA Notes:
N/A

:Issues Addressed:
N/A

Reviewed-on: http://gpro.lge.com/c/webosose/webos-fluentbit-plugins/+/338206
Tested-by: Myungchul Kim <myungchul.kim@lge.com>

+++ plugins/common/util/MSGPackUtil.h
@@ -35,7 +35,7 @@ public:
     static void putValue(msgpack_packer* packer, const string& key, const JValue& value);
     static void putValue(msgpack_packer* packer, const string& key, const string& value);
     static void putValue(msgpack_packer* packer, const string& key, const char* value);
-    static void putValue(msgpack_packer* packer, const string& key, int value);
+    static void putValue(msgpack_packer* packer, const string& key, int64_t value);
     static void putValue(msgpack_packer* packer, const string& key, double value);
     static void putValue(msgpack_packer* packer, const string& key, bool value);
